idalib_sys

Module bytes

Functions

get_flags⚠

Get flags value for address ‘ea’. The byte value is not included in the flags. This function should be used if the operand types of any operand beyond the first two operands is required. This function is more expensive to use than get_flags32() \return 0 if address is not present in the program

idalib_get_byte⚠

idalib_get_bytes⚠

idalib_get_dword⚠

idalib_get_qword⚠

idalib_get_word⚠

is_byte⚠

is_code⚠

Does flag denote start of an instruction?

is_data⚠

Does flag denote start of data?

is_double⚠

is_dword⚠

is_float⚠

is_off⚠

is_oword⚠

is_qword⚠

is_strlit⚠

is_word⚠

Type Aliases

flags64_t