idalib_sys::ffi::bindgen
Module root

Re-exports

pub use cxxbridge::gdl_graph_t;

pub use cxxbridge::qstring_char_AutocxxConcrete;

pub use cxxbridge::qrefcnt_t_cfunc_t_AutocxxConcrete;

pub use super::super::insn_t;

pub use cxxbridge::reg_finder_t;

pub use super::super::cexpr_t;

pub use super::super::cfunc_t;

Structs

_IO_FILE

_qstring

Reimplementation of the string class from STL. Only the most essential functions are implemented. The reason why we have this is because it is not compiler dependent (hopefully) and therefore can be used in IDA API

bitrange_t

This class manages the offset and size of a value that occupies a number of contiguous bits within some container - generally a byte array. A special state - empty range (both offset and size are zeroes) - determines the value as all bits of the container.

carg_t

Function argument

carglist_t

Function argument list

fpvalue_t

\file ieee.h

func_t

A function is a set of continuous ranges of addresses with characteristics

hexrays_failure_t

Exception object: decompiler failure information

lock_func

Helper class to lock a function pointer so it stays valid

lock_segment

Helper class to lock a segment pointer so it stays valid

node_iterator

\file gdl.hpp

plugin_t

A plugin is a module in the plugins subdirectory that can perform an action asked by the user. (usually via pressing a hotkey)

processor_t

Describes a processor module (IDP). An IDP file may have only one such structure called LPH. The kernel will copy it to dbctx_t::ph.

qbasic_block_t

Information about a basic block of a \ref qflow_chart_t

qflow_chart_t

A flow chart for a function, or a set of address ranges

qrefcnt_t

Smart pointer to objects derived from ::qrefcnt_obj_t

range_t

\file lines.hpp

rangevec_t

segment_t

Describes a program segment

tinfo_t

Primary mechanism for managing type information

vc_printer_t

Helper class to convert cfunc_t into text.

xrefblk_t

Structure to enumerate all xrefs. This structure provides a way to access cross-references from a given address. For example:

Enums

cref_t

CODE xref types

dref_t

DATA xref types

fc_block_type_t

Flow chart block types

filetype_t

Known input file formats (kept in \inf{filetype}):

fpvalue_error_t

\defgroup REAL_ERROR_ Floating point/IEEE Conversion codes Return values for ieee_realcvt and processor_t::realcvt request

merror_t

\defgroup MERR_ Microcode error codes @{

Constants

FC_APPND

FC_CALL_ENDS

FC_CHKBREAK

FC_NOEXT

FC_NOPREDS

FC_OUTLINES

FC_PRINT

FC_RESERVED

FUNC_BOTTOMBP

FUNC_FAR

FUNC_FRAME

FUNC_FUZZY_SP

FUNC_HIDDEN

FUNC_LIB

FUNC_LUMINA

FUNC_NORET

FUNC_NORET_PENDING

FUNC_OUTLINE

FUNC_PROLOG_OK

FUNC_PURGED_OK

FUNC_REANALYZE

FUNC_RESERVED

FUNC_SP_READY

FUNC_STATICDEF

FUNC_TAIL

FUNC_THUNK

FUNC_USERFAR

IRI_EXTENDED

IRI_RET_LITERALLY

IRI_SKIP_RETTARGET

IRI_STRICT

PLUGIN_DBG

PLUGIN_DRAW

PLUGIN_FIX

PLUGIN_HIDE

PLUGIN_MOD

PLUGIN_MULTI

PLUGIN_PROC

PLUGIN_SCRIPTED

PLUGIN_SEG

PLUGIN_UNL

SEGPERM_EXEC

SEGPERM_MAXVAL

SEGPERM_READ

SEGPERM_WRITE

SEG_ABSSYM

SEG_BSS

SEG_CODE

SEG_COMM

SEG_DATA

SEG_GRP

SEG_IMEM

SEG_IMP

SEG_MAX_SEGTYPE_CODE

SEG_NORM

SEG_NULL

SEG_UNDF

SEG_XTRN

XREF_ALL

XREF_BASE

XREF_DATA

XREF_FAR

XREF_MASK

XREF_PASTEND

XREF_TAIL

XREF_USER

dt_bitfild

dt_byte

dt_byte16

dt_byte32

dt_byte64

dt_code

dt_double

dt_dword

dt_float

dt_fword

dt_half

dt_ldbl

dt_packreal

dt_qword

dt_string

dt_tbyte

dt_unicode

dt_void

dt_word

o_displ

< Memory Ref [Base Reg + Index Reg + Displacement].

o_far

< Immediate Far Address (CODE).

o_idpspec0

< processor specific type.

o_idpspec1

< processor specific type.

o_idpspec2

< processor specific type.

o_idpspec3

< processor specific type.

o_idpspec4

< processor specific type.

o_idpspec5

< processor specific type. < (there can be more processor specific types)

o_imm

< Immediate Value.

o_mem

< Direct Memory Reference (DATA).

o_near

< Immediate Near Address (CODE).

o_phrase

< Memory Ref [Base Reg + Index Reg].

o_reg

< General Register (al,ax,es,ds…).

o_void

< No Operand.

saAbs

saGroup

saRel4K

saRel32Bytes

saRel64Bytes

saRel128Bytes

saRel512Bytes

saRel1024Bytes

saRel2048Bytes

saRelByte

saRelDble

saRelPage

saRelPara

saRelQword

saRelWord

saRel_MAX_ALIGN_CODE

Type Aliases

FILE

asize_t

bgcolor_t

cfuncptr_t

Smart pointer to objects derived from ::qrefcnt_obj_t

ea_t

flags64_t

optype_t

\defgroup operands Operands Work with instruction operands

qstring

Reimplementation of the string class from STL. Only the most essential functions are implemented. The reason why we have this is because it is not compiler dependent (hopefully) and therefore can be used in IDA API

uchar

uint16

uint32

uint64

uval_t