idalib_sys::ffi::bindgen::root
Struct bitrange_t

#[repr(C, align(2))]
pub struct bitrange_t {
    _pinned: PhantomData<PhantomPinned>,
    _non_send_sync: PhantomData<[*const u8; 0]>,
    _data: UnsafeCell<MaybeUninit<[u8; 4]>>,
}

This class manages the offset and size of a value that occupies a number of contiguous bits within some container - generally a byte array. A special state - empty range (both offset and size are zeroes) - determines the value as all bits of the container.

Fields
_pinned: PhantomData<PhantomPinned>_non_send_sync: PhantomData<[*const u8; 0]>_data: UnsafeCell<MaybeUninit<[u8; 4]>>