idalib_sys::ffi
Module cxxbridge

Structs

gdl_graph_t

gdl graph interface - includes only functions required to draw it

qrefcnt_t_cfunc_t_AutocxxConcrete

qstring_char_AutocxxConcrete

reg_finder_t

Functions

_IO_FILE_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

_IO_FILE_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

_IO_FILE_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

_IO_FILE_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

append_cmt⚠

Append to an indented comment. Creates a new comment if none exists. Appends a newline character and the specified string otherwise. \param ea linear address \param str comment string to append \param rptble append to repeatable comment? \return success

auto_queue_empty_autocxx_wrapper_0xdde2bac051107010⚠

auto_wait⚠

Process everything in the queues and return true. \return false if the user clicked cancel. (the wait box must be displayed by the caller if desired)

autocxx_make_string_0xdde2bac051107010

bitrange_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

bitrange_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

bitrange_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

bitrange_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

build_strlist⚠

Rebuild the string list.

calc_thunk_func_target⚠

Calculate target of a thunk function. \param pfn pointer to function (may not be nullptr) \param fptr out: will hold address of a function pointer (if indirect jump) \return the target function or #BADADDR

carg_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

carg_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

carg_t_compare_autocxx_wrapper_0xdde2bac051107010⚠

carg_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

carglist_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_compare_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_new1_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_print1_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_print_autocxx_wrapper_0xdde2bac051107010⚠

carglist_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

cast_func_t_to_range_t_autocxx_wrapper_0xdde2bac051107010

cast_qbasic_block_t_to_range_t_autocxx_wrapper_0xdde2bac051107010

cast_segment_t_to_range_t_autocxx_wrapper_0xdde2bac051107010

clear_strlist⚠

Clear the string list.

close_database⚠

\brief Close the current database \param save boolean value, save or discard changes

decode_insn⚠

Analyze the specified address and fill ‘out’. This function does not modify the database. It just tries to interpret the specified address as an instruction and fills the ‘out’ structure. \param out the resulting instruction \param ea linear address \return the length of the (possible) instruction or 0

desc_autocxx_wrapper_0xdde2bac051107010⚠

enable_console_messages⚠

\brief Enable console messages \param enable flag to activate or deactivate console messages by default, console messages are disabled

entry_autocxx_wrapper_0xdde2bac051107010⚠

exists_autocxx_wrapper_0xdde2bac051107010⚠

exit_autocxx_wrapper_0xdde2bac051107010⚠

find_plugin⚠

Find a user-defined plugin and optionally load it. \param name short plugin name without path and extension, or absolute path to the file name \param load_if_needed if the plugin is not present in the memory, try to load it \return pointer to plugin description block

fpvalue_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

fpvalue_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

fpvalue_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

fpvalue_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

func_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

func_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

func_t_compare_autocxx_wrapper_0xdde2bac051107010⚠

func_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

func_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

func_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

func_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

gdl_graph_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_begin_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_destructor_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_empty_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_end_autocxx_wrapper_0xdde2bac051107010⚠

gdl_graph_t_size_autocxx_wrapper_0xdde2bac051107010⚠

get_code16_mode_autocxx_wrapper_0xdde2bac051107010⚠

get_edge_color_autocxx_wrapper_0xdde2bac051107010⚠

get_entry⚠

Get entry point address by its ordinal \param ord ordinal number of entry point \return address or #BADADDR

get_entry_ordinal⚠

Get ordinal number of an entry point. \param idx internal number of entry point. Should be in the range 0..get_entry_qty()-1 \return ordinal number or 0.

get_entry_qty⚠

Get number of entry points

get_func⚠

Get pointer to function structure by address. \param ea any address in a function \return ptr to a function or nullptr. This function returns a function entry chunk.

get_func_num⚠

Get ordinal number of a function. \param ea any address in the function \return number of function (0..get_func_qty()-1). -1 means ‘no function at the specified address’.

get_func_qty⚠

Get total number of functions in the program

get_node_color_autocxx_wrapper_0xdde2bac051107010⚠

get_node_label_autocxx_wrapper_0xdde2bac051107010⚠

get_ph⚠

get_reg_info_autocxx_wrapper_0xdde2bac051107010⚠

Get register information - useful for registers like al, ah, dil, etc. Example: this function for “al” returns “eax” in 32bit mode \return main register name (nullptr no such register)

get_regfinder_autocxx_wrapper_0xdde2bac051107010⚠

get_segm_by_name⚠

Get pointer to segment by its name. If there are several segments with the same name, returns the first of them. \param name segment name. may be nullptr. \return nullptr or pointer to segment structure

get_segm_qty⚠

Get number of segments

get_strlist_qty⚠

Get number of elements in the string list. The list will be loaded from the database (if saved) or built from scratch.

getn_func⚠

Get pointer to function structure by number. \param n number of function, is in range 0..get_func_qty()-1 \return ptr to a function or nullptr. This function returns a function entry chunk.

getnseg⚠

Get pointer to segment by its number. \warning Obsoleted because it can slow down the debugger (it has to refresh the whole memory segmentation to calculate the correct answer) \param n segment number in the range (0..get_segm_qty()-1) \return nullptr or pointer to segment structure

getseg⚠

Get pointer to segment by linear address. \param ea linear address belonging to the segment \return nullptr or pointer to segment structure

has_external_refs⚠

Does ‘ea’ have references from outside of ‘pfn’?

hexrays_failure_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

hexrays_failure_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

hexrays_failure_t_new1_autocxx_wrapper_0xdde2bac051107010⚠

hexrays_failure_t_new2_autocxx_wrapper_0xdde2bac051107010⚠

hexrays_failure_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

hexrays_failure_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

init_hexrays_plugin⚠

Check that your plugin is compatible with hex-rays decompiler. This function must be called before calling any other decompiler function. \param flags reserved, must be 0 \return true if the decompiler exists and is compatible with your plugin

is_align_insn_autocxx1⚠

If the instruction at ‘ea’ looks like an alignment instruction, return its length in bytes. Otherwise return 0.

is_basic_block_end⚠

Is the instruction the end of a basic block?

is_call_insn⚠

Is the instruction a “call”?

is_indirect_jump_insn⚠

Is the instruction an indirect jump?

is_ret_insn⚠

@}

lock_func_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

lock_func_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

lock_func_destructor_autocxx_wrapper_0xdde2bac051107010⚠

lock_func_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

lock_func_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

lock_segment_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

lock_segment_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

lock_segment_destructor_autocxx_wrapper_0xdde2bac051107010⚠

lock_segment_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

lock_segment_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

make_signatures⚠

\brief Generate .sig and .pat files for the current database \param only_pat Generate .pat file only \return true in case of success

next_head⚠

Get start of next defined item. \param ea begin search at this address \param maxea not included in the search range \return #BADADDR if none exists.

node_iterator_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

node_iterator_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

node_iterator_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

node_iterator_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

node_qty_autocxx_wrapper_0xdde2bac051107010⚠

npred_autocxx_wrapper_0xdde2bac051107010⚠

nsucc_autocxx_wrapper_0xdde2bac051107010⚠

open_database⚠

\brief Open the database specified in file_path argument If the database did not exist, a new database will be created and the input file will be loaded Note: All library functions must be called from the same thread that initialized the library The library is single-threaded, and performing database operations from a different thread than the initialization thread may lead to undefined behavior \param file_path the file name to be loaded \param run_auto if set to true, library will run also auto analysis \param args optional arguments, respecting IDA’s command-line arguments format \return 0 if successfully opened, otherwise error code

plugin_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

plugin_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

plugin_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

plugin_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

plugin_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

pred_autocxx_wrapper_0xdde2bac051107010⚠

prev_head⚠

Get start of previous defined item. \param ea begin search at this address \param minea included in the search range \return #BADADDR if none exists.

print_edge_autocxx_wrapper_0xdde2bac051107010⚠

print_graph_attributes_autocxx_wrapper_0xdde2bac051107010⚠

print_node_attributes_autocxx_wrapper_0xdde2bac051107010⚠

print_node_autocxx_wrapper_0xdde2bac051107010⚠

processor_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

processor_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

processor_t_get_canon_feature_autocxx_wrapper_0xdde2bac051107010⚠

Get the instruction features (combination of \ref CF_)

processor_t_get_canon_mnem_autocxx_wrapper_0xdde2bac051107010⚠

Get the instruction name

processor_t_is_canon_insn_autocxx_wrapper_0xdde2bac051107010⚠

Does the given value specify a valid instruction for this instruction set?. See #instruc_start and #instruc_end

processor_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

qbasic_block_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

qbasic_block_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

qbasic_block_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

qflow_chart_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_create1_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_create_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_destructor_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_get_node_label_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_new1_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_npred_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_nsucc_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_pred_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_print_node_attributes_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_size_autocxx_wrapper_0xdde2bac051107010⚠

qflow_chart_t_succ_autocxx_wrapper_0xdde2bac051107010⚠

range_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

range_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

range_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

range_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

range_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

range_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

rangevec_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

rangevec_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

realcvt_autocxx_wrapper_0xdde2bac051107010⚠

run_plugin⚠

Run a loaded plugin with the specified argument. \param ptr pointer to plugin description block \param arg argument to run with

segment_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

segment_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

segment_t_compare_autocxx_wrapper_0xdde2bac051107010⚠

segment_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

Constructor

segment_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

segment_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

segment_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

set_cmt⚠

Set an indented comment. \param ea linear address \param comm comment string - nullptr: do nothing (return 0) - “” : delete comment \param rptble is repeatable? \return success

set_idp_options_autocxx_wrapper_0xdde2bac051107010⚠

set_screen_ea⚠

\brief Set screen ea, let the user specify the current screen ea subsequent calls to get_screen_ea will return this value Please note that with idalib there is no screen, this is provided just for being able to use the get_screen_ea \param ea screen ea

str2reg_autocxx1⚠

succ_autocxx_wrapper_0xdde2bac051107010⚠

term_hexrays_plugin⚠

Stop working with hex-rays decompiler.

tinfo_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

tinfo_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

tinfo_t_new3_autocxx_wrapper_0xdde2bac051107010⚠

Constructor

vc_printer_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

vc_printer_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

xrefblk_t_autocxx_alloc_autocxx_wrapper_0xdde2bac051107010⚠

xrefblk_t_autocxx_free_autocxx_wrapper_0xdde2bac051107010⚠

xrefblk_t_first_from⚠

\name Helper functions Should not be called directly! @{

xrefblk_t_first_to⚠

xrefblk_t_new_autocxx_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized default constructor.

xrefblk_t_new_synthetic_const_copy_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized copy constructor.

xrefblk_t_new_synthetic_move_ctor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized move constructor.

xrefblk_t_next_from⚠

xrefblk_t_next_to⚠

xrefblk_t_synthetic_destructor_0xdde2bac051107010_autocxx_wrapper_0xdde2bac051107010⚠

Synthesized destructor.

Type Aliases

_IO_FILE

bitrange_t

This class manages the offset and size of a value that occupies a number of contiguous bits within some container - generally a byte array. A special state - empty range (both offset and size are zeroes) - determines the value as all bits of the container.

c_int

c_uint

c_ulonglong

c_ushort

c_void

carg_t

Function argument

carglist_t

Function argument list

cexpr_t

cfunc_t

cref_t

CODE xref types

dref_t

DATA xref types

fc_block_type_t

Flow chart block types

filetype_t

Known input file formats (kept in \inf{filetype}):

fpvalue_error_t

\defgroup REAL_ERROR_ Floating point/IEEE Conversion codes Return values for ieee_realcvt and processor_t::realcvt request

fpvalue_t

\file ieee.h

func_t

A function is a set of continuous ranges of addresses with characteristics

hexrays_failure_t

Exception object: decompiler failure information

insn_t

lock_func

Helper class to lock a function pointer so it stays valid

lock_segment

Helper class to lock a segment pointer so it stays valid

merror_t

\defgroup MERR_ Microcode error codes @{

node_iterator

\file gdl.hpp

plugin_t

A plugin is a module in the plugins subdirectory that can perform an action asked by the user. (usually via pressing a hotkey)

processor_t

Describes a processor module (IDP). An IDP file may have only one such structure called LPH. The kernel will copy it to dbctx_t::ph.

qbasic_block_t

Information about a basic block of a \ref qflow_chart_t

qflow_chart_t

A flow chart for a function, or a set of address ranges

range_t

\file lines.hpp

rangevec_t

segment_t

Describes a program segment

tinfo_t

Primary mechanism for managing type information

vc_printer_t

Helper class to convert cfunc_t into text.

xrefblk_t

Structure to enumerate all xrefs. This structure provides a way to access cross-references from a given address. For example: