Type Alias idalib_sys::ffi::cxxbridge::merror_t

pub type merror_t = merror_t;

\defgroup MERR_ Microcode error codes @{

Aliased Type

enum merror_t {

    MERR_OK = 0,
    MERR_BLOCK = 1,
    MERR_INTERR = -1,
    MERR_INSN = -2,
    MERR_MEM = -3,
    MERR_BADBLK = -4,
    MERR_BADSP = -5,
    MERR_PROLOG = -6,
    MERR_SWITCH = -7,
    MERR_EXCEPTION = -8,
    MERR_HUGESTACK = -9,
    MERR_LVARS = -10,
    MERR_BITNESS = -11,
    MERR_BADCALL = -12,
    MERR_BADFRAME = -13,
    MERR_UNKTYPE = -14,
    MERR_BADIDB = -15,
    MERR_SIZEOF = -16,
    MERR_REDO = -17,
    MERR_CANCELED = -18,
    MERR_RECDEPTH = -19,
    MERR_OVERLAP = -20,
    MERR_PARTINIT = -21,
    MERR_COMPLEX = -22,
    MERR_LICENSE = -23,
    MERR_ONLY32 = -24,
    MERR_ONLY64 = -25,
    MERR_BUSY = -26,
    MERR_FARPTR = -27,
    MERR_EXTERN = -28,
    MERR_FUNCSIZE = -29,
    MERR_BADRANGES = -30,
    MERR_BADARCH = -31,
    MERR_DSLOT = -32,
    MERR_STOP = -33,
    MERR_CLOUD = -34,
    MERR_MAX_ERR = 34,
    MERR_LOOP = -35,

}

Variants

MERR_OK = 0

< ok

MERR_BLOCK = 1

< no error, switch to new block

MERR_INTERR = -1

< internal error

MERR_INSN = -2

< cannot convert to microcode

MERR_MEM = -3

< not enough memory

MERR_BADBLK = -4

< bad block found

MERR_BADSP = -5

< positive sp value has been found

MERR_PROLOG = -6

< prolog analysis failed

MERR_SWITCH = -7

< wrong switch idiom

MERR_EXCEPTION = -8

< exception analysis failed

MERR_HUGESTACK = -9

< stack frame is too big

MERR_LVARS = -10

< local variable allocation failed

MERR_BITNESS = -11

< 16-bit functions cannot be decompiled

MERR_BADCALL = -12

< could not determine call arguments

MERR_BADFRAME = -13

< function frame is wrong

MERR_UNKTYPE = -14

< undefined type %s (currently unused error code)

MERR_BADIDB = -15

< inconsistent database information

MERR_SIZEOF = -16

< wrong basic type sizes in compiler settings

MERR_REDO = -17

< redecompilation has been requested

MERR_CANCELED = -18

< decompilation has been cancelled

MERR_RECDEPTH = -19

< max recursion depth reached during lvar allocation

MERR_OVERLAP = -20

< variables would overlap: %s

MERR_PARTINIT = -21

< partially initialized variable %s

MERR_COMPLEX = -22

< too complex function

MERR_LICENSE = -23

< no license available

MERR_ONLY32 = -24

< only 32-bit functions can be decompiled for the current database

MERR_ONLY64 = -25

< only 64-bit functions can be decompiled for the current database

MERR_BUSY = -26

< already decompiling a function

MERR_FARPTR = -27

< far memory model is supported only for pc

MERR_EXTERN = -28

< special segments cannot be decompiled

MERR_FUNCSIZE = -29

< too big function

MERR_BADRANGES = -30

< bad input ranges

MERR_BADARCH = -31

< current architecture is not supported

MERR_DSLOT = -32

< bad instruction in the delay slot

MERR_STOP = -33

< no error, stop the analysis

MERR_CLOUD = -34

< cloud: %s

MERR_MAX_ERR = 34

MERR_LOOP = -35

< internal code: redo last loop (never reported)