idalib_sys

Module func

Modules

cfg_flags

flags

Structs

func_t

A function is a set of continuous ranges of addresses with characteristics

func_tail_t

gdl_graph_t

gdl graph interface - includes only functions required to draw it

lock_func

Helper class to lock a function pointer so it stays valid \deprecated Use lock_func_ea instead.

qbasic_block_t

Information about a basic block of a \ref qflow_chart_t

qflow_chart_t

\deprecated Use qflow_chart_ea_t for safer ea-based access.

Enums

fc_block_type_t

Flow chart block types

Functions

add_func⚠

Add a new function. If the function end address is #BADADDR, then IDA will try to determine the function bounds by calling find_func_bounds(…, #FIND_FUNC_DEFINE). \param ea1 start address \param ea2 end address \return success

calc_thunk_func_target⚠

Calculate target of a thunk function. \param pfn pointer to function (may not be nullptr) \param fptr out: will hold address of a function pointer (if indirect jump) \return the target function or #BADADDR \deprecated Use calc_thunk_function_target() for safer access.

del_func⚠

Delete a function. \param ea any address in the function entry chunk \return success

get_func⚠

Get pointer to function structure by address. \param ea any address in a function \return ptr to a function or nullptr. This function returns a function entry chunk. \deprecated Use get_func_start() or get_func_entry_info() for safer access.

get_func_num⚠

Get ordinal number of a function. \param ea any address in the function \return number of function (0..get_func_qty()-1). -1 means ‘no function at the specified address’.

get_func_qty⚠

Get total number of functions in the program

getn_func⚠

Get pointer to function structure by number. \param n number of function, is in range 0..get_func_qty()-1 \return ptr to a function or nullptr. This function returns a function entry chunk. \deprecated Use get_func_ea_by_num() or get_func_entry_info_by_num() for safer access.

idalib_func_flags⚠

idalib_func_flow_chart⚠

idalib_func_name⚠

idalib_func_tails⚠

idalib_get_func_cmt⚠

idalib_qbasic_block_preds⚠

idalib_qbasic_block_succs⚠

idalib_qflow_graph_getn_block⚠

idalib_set_func_cmt⚠