idalib_sys

Module hexrays

Structs

HexRaysError

carg_t

Function argument

carglist_t

Function argument list

cblock_iter

cblock_t

Compound statement (curly braces)

cexpr_t

Ctree item: expression. Depending on the exact expression item type, various fields of this structure are used.

cfunc_t

Decompiled function. Decompilation result is kept here.

cfuncptr_t

cinsn_t

Ctree item: statement. Depending on the exact statement type, various fields of the union are used.

citem_t

Basic ctree item. This is an abstract class (but we don’t use virtual functions in ctree, so the compiler will not disallow you to create citem_t instances). However, items of pure citem_t type must never be created. Two classes, cexpr_t and cinsn_t are derived from it.

cswitch_t

Switch statement

cthrow_t

Throw statement

ctry_t

C++ Try statement. This structure is also used to represent wind statements.

Enums

HexRaysErrorCode

Functions

change_hexrays_config⚠

Parse DIRECTIVE and update the current configuration variables. For the syntax see hexrays.cfg

decompile_function⚠

idalib_hexrays_cblock_iter⚠

idalib_hexrays_cblock_iter_next⚠

idalib_hexrays_cblock_len⚠

idalib_hexrays_cfunc_body⚠

idalib_hexrays_cfunc_pseudocode⚠

idalib_hexrays_cfuncptr_inner⚠

idalib_hexrays_decompile_function⚠

init_hexrays_plugin⚠

Check that your plugin is compatible with hex-rays decompiler. This function must be called before calling any other decompiler function. \param flags reserved, must be 0 \return true if the decompiler exists and is compatible with your plugin

term_hexrays_plugin⚠

Stop working with hex-rays decompiler.