idalib_sys::hexrays
Module __impl

Structs

__IncompleteArrayField

_qstring

Reimplementation of the string class from STL. Only the most essential functions are implemented. The reason why we have this is because it is not compiler dependent (hopefully) and therefore can be used in IDA API

argloc_t

Describes an argument location. A typical argument is stored in one location, either a register or a stack slot. \n However, some arguments can be stored in multiple locations, for example in a pair \n of registers. In some really complex cases an argument can be located in multiple \n registers and some stack slots. This class can describe all these cases.

argpart_t

Subsection of an argument location

array_parameters_t

\name Array representation @{ Describes how to display an array

bitset_t

Bit set class. See https://en.wikipedia.org/wiki/Bit_array

bitset_t_iterator

bytevec_t

Vector of bytes (use for dynamic memory)

carglist_t

Function argument list

casm_t

asm statement

catchexpr_t

Catch expression

cblock_t

Compound statement (curly braces)

ccase_t

Switch case. Usually cinsn_t is a block

ccases_t

Vector of switch cases

ccatch_t

Catch clause: “catch ( type obj )”

cdo_t

Do-loop

ceinsn_t

Statement with an expression. This is a base class for various statements with expressions.

cexpr_t

Ctree item: expression. Depending on the exact expression item type, various fields of this structure are used.

cexpr_t__bindgen_ty_1__bindgen_ty_1

cexpr_t__bindgen_ty_1__bindgen_ty_2

cfor_t

For-loop

cfunc_t

Decompiled function. Decompilation result is kept here.

cgoto_t

Goto statement

cif_t

If statement

cinsn_t

Ctree item: statement. Depending on the exact statement type, various fields of the union are used.

citem_cmt_t

Ctree item comment. For each comment we remember its body and the fact of its retrieval

citem_locator_t

Generic ctree item locator. It can be used for instructions and some expression types. However, we need more precise locators for other items (e.g. for numbers)

citem_t

Basic ctree item. This is an abstract class (but we don’t use virtual functions in ctree, so the compiler will not disallow you to create citem_t instances). However, items of pure citem_t type must never be created. Two classes, cexpr_t and cinsn_t are derived from it.

cloop_t

Base class for loop statements

cnumber_t

An immediate number

creturn_t

Return statement

cswitch_t

Switch statement

cthrow_t

Throw statement

ctry_t

C++ Try statement. This structure is also used to represent wind statements.

custom_data_type_info_t

cwhile_t

While-loop

fnumber_t

Floating point constant. Used for mop_fn For more details, please see the ieee.h file from IDA SDK.

fpvalue_t

\file ieee.h

func_t

A function is a set of continuous ranges of addresses with characteristics

func_t__bindgen_ty_1__bindgen_ty_1

attributes of a function entry chunk

func_t__bindgen_ty_1__bindgen_ty_2

attributes of a function tail chunk

hexwarn_t

Warning instances

ivl_t

ivl_tpl

ivl_with_name_t

ivlset_t

ivlset_tpl

llabel_t

lvar_locator_t

Local variable locator. Local variables are located using definition ea and location. Each variable must have a unique locator, this is how we tell them apart.

lvar_ref_t

Reference to a local variable. Used by mop_l

lvar_t

Definition of a local variable (register or stack) #var #lvar

lvars_t

Vector of local variables

mba_ranges_t

Ranges to decompile. Either a function or an explicit vector of ranges.

mba_t

Array of micro blocks representing microcode for a decompiled function. The first micro block is the entry point, the last one is the exit point. The entry and exit blocks are always empty. The exit block is generated at MMAT_LOCOPT maturity level.

mblock_t

Microcode of one basic block. All blocks are part of a doubly linked list. They can also be addressed by indexing the mba->natural array. A block contains a doubly linked list of instructions, various location lists that are used for data flow analysis, and other attributes.

mblock_t__bindgen_vtable

mcallarg_t

A call argument

mcallinfo_t

Information about a call

mcases_t

List of switch cases and targets

minsn_t

Microinstruction class #insn

mlist_t

mnumber_t

An integer constant. Used for mop_n We support 64-bit values but 128-bit values can be represented with mop_p

mop_addr_t

Address of an operand (mop_l, mop_v, mop_S, mop_r)

mop_pair_t

Pair of operands

mop_t

A microinstruction operand. This is the smallest building block of our microcode. Operands will be part of instructions, which are then grouped into basic blocks. The microcode consists of an array of such basic blocks + some additional info.

number_format_t

Number representation. This structure holds information about a number format.

operand_locator_t

Operand locator. It is used to denote a particular operand in the ctree, for example, when the user right clicks on a constant and requests to represent it, say, as a hexadecimal number.

qlist

Linked list Note: linked list is not movable!

qlist_const_iterator

qlist_const_reverse_iterator

qlist_datanode_t

qlist_iterator

qlist_listnode_t

qlist_reverse_iterator

qvector

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

range_t

\file lines.hpp

rangeset_t

An ordered set of non-overlapping address ranges

rangevec_t

refinfo_t

Information about a reference

reg_info_t

Get register number and size from register name

regarg_t

Register argument description. regargs are destroyed when the full function type is determined.

regvar_t

rlist_t

rrel_t

Register-relative argument location

scattered_aloc_t

Used to manage arguments that are described by multiple locations (also see ::ALOC_DIST)

scif_t

Scattered operand info. Used for mop_sc

simpleline_t

Maintain basic information for a line in a custom view

std_allocator

std_allocator_rebind

std_less

std_map

std_map_value_compare

std_pair

std_pair___zero_as_null_pointer_constant

stkpnt_t

\file funcs.hpp

stkvar_ref_t

Reference to a stack variable. Used for mop_S

tinfo_t

Primary mechanism for managing type information

treeloc_t

Ctree location. Used to denote comment locations.

type_attr_t

\addtogroup tattr_ext Extended type attributes @{ Extended type attributes.

udm_t

An object to represent struct or union members

value_repr_t

Visual representation of a member of a complex type (struct/union/enum)

value_repr_t__bindgen_ty_1__bindgen_ty_1

var_ref_t

Reference to a local variable

vc_printer_t

Helper class to convert cfunc_t into text.

vd_printer_t

Base helper class to convert binary data structures into text. Other classes are derived from this class.

vd_printer_t__bindgen_vtable

vdloc_t

@}

Constants

CV_FAST

CV_INSNS

CV_PARENTS

CV_POST

CV_PRUNE

CV_RESTART

DECOMP_ALL_BLKS

DECOMP_GXREFS_DEFLT

DECOMP_GXREFS_FORCE

DECOMP_GXREFS_NOUPD

DECOMP_NO_CACHE

DECOMP_NO_FRAME

DECOMP_NO_HIDE

DECOMP_NO_WAIT

DECOMP_NO_XREFS

DECOMP_VOID_MBA

DECOMP_WARNINGS

ctree_maturity_t_CMAT_BUILT

< just generated

ctree_maturity_t_CMAT_CASTED

< added necessary casts

ctree_maturity_t_CMAT_CPA

< corrected pointer arithmetic

ctree_maturity_t_CMAT_FINAL

< ready-to-use

ctree_maturity_t_CMAT_NICE

< nicefied expressions

ctree_maturity_t_CMAT_TRANS1

< applied first wave of transformations

ctree_maturity_t_CMAT_TRANS2

< applied second wave of transformations

ctree_maturity_t_CMAT_TRANS3

< applied third wave of transformations

ctree_maturity_t_CMAT_ZERO

< does not exist

ctype_t_cit_asm

< asm-statement

ctype_t_cit_block

< block-statement: { … }

ctype_t_cit_break

< break-statement

ctype_t_cit_continue

< continue-statement

ctype_t_cit_do

< do-statement

ctype_t_cit_empty

< instruction types start here

ctype_t_cit_end

ctype_t_cit_expr

< expression-statement: expr;

ctype_t_cit_for

< for-statement

ctype_t_cit_goto

< goto-statement

ctype_t_cit_if

< if-statement

ctype_t_cit_return

< return-statement

ctype_t_cit_switch

< switch-statement

ctype_t_cit_throw

< C++ throw-statement

ctype_t_cit_try

< C++ try-statement

ctype_t_cit_while

< while-statement

ctype_t_cot_add

< x + y

ctype_t_cot_asg

< x = y

ctype_t_cot_asgadd

< x += y

ctype_t_cot_asgband

< x &= y

ctype_t_cot_asgbor

< x |= y

ctype_t_cot_asgmul

< x *= y

ctype_t_cot_asgsdiv

< x /= y signed

ctype_t_cot_asgshl

< x <<= y

ctype_t_cot_asgsmod

< x %= y signed

ctype_t_cot_asgsshr

< x >>= y signed

ctype_t_cot_asgsub

< x -= y

ctype_t_cot_asgudiv

< x /= y unsigned

ctype_t_cot_asgumod

< x %= y unsigned

ctype_t_cot_asgushr

< x >>= y unsigned

ctype_t_cot_asgxor

< x ^= y

ctype_t_cot_band

< x & y

ctype_t_cot_bnot

< ~x

ctype_t_cot_bor

< x | y

ctype_t_cot_call

< x(…)

ctype_t_cot_cast

< (type)x

ctype_t_cot_comma

< x, y

ctype_t_cot_empty

ctype_t_cot_eq

< x == y int or fpu (see EXFL_FPOP)

ctype_t_cot_fadd

< x + y fp

ctype_t_cot_fdiv

< x / y fp

ctype_t_cot_fmul

< x * y fp

ctype_t_cot_fneg

< -x fp

ctype_t_cot_fnum

< fpc

ctype_t_cot_fsub

< x - y fp

ctype_t_cot_helper

< arbitrary name

ctype_t_cot_idx

< x[y]

ctype_t_cot_insn

< instruction in expression, internal representation only

ctype_t_cot_land

< x && y

ctype_t_cot_last

ctype_t_cot_lnot

< !x

ctype_t_cot_lor

< x || y

ctype_t_cot_memptr

< x->m, access size in ‘ptrsize’

ctype_t_cot_memref

< x.m

ctype_t_cot_mul

< x * y

ctype_t_cot_ne

< x != y int or fpu (see EXFL_FPOP)

ctype_t_cot_neg

< -x

ctype_t_cot_num

< n

ctype_t_cot_obj

< obj_ea

ctype_t_cot_postdec

< x–

ctype_t_cot_postinc

< x++

ctype_t_cot_predec

< –x

ctype_t_cot_preinc

< ++x

ctype_t_cot_ptr

< *x, access size in ‘ptrsize’

ctype_t_cot_ref

< &x

ctype_t_cot_sdiv

< x / y signed

ctype_t_cot_sge

< x >= y signed or fpu (see EXFL_FPOP)

ctype_t_cot_sgt

< x > y signed or fpu (see EXFL_FPOP)

ctype_t_cot_shl

< x << y

ctype_t_cot_sizeof

< sizeof(x)

ctype_t_cot_sle

< x <= y signed or fpu (see EXFL_FPOP)

ctype_t_cot_slt

< x < y signed or fpu (see EXFL_FPOP)

ctype_t_cot_smod

< x % y signed

ctype_t_cot_sshr

< x >> y signed

ctype_t_cot_str

< string constant (user representation)

ctype_t_cot_sub

< x - y

ctype_t_cot_tern

< x ? y : z

ctype_t_cot_type

< arbitrary type

ctype_t_cot_udiv

< x / y unsigned

ctype_t_cot_uge

< x >= y unsigned

ctype_t_cot_ugt

< x > y unsigned

ctype_t_cot_ule

< x <= y unsigned

ctype_t_cot_ult

< x < y unsigned

ctype_t_cot_umod

< x % y unsigned

ctype_t_cot_ushr

< x >> y unsigned

ctype_t_cot_var

< v

ctype_t_cot_xor

< x ^ y

funcrole_t_ROLE_3WAYCMP0

< 3-way compare helper, returns -1/0/1

funcrole_t_ROLE_3WAYCMP1

< 3-way compare helper, returns 0/1/2

funcrole_t_ROLE_ABS

< integer absolute value

funcrole_t_ROLE_ALLOCA

< alloca() function

funcrole_t_ROLE_BITTEST

< [lock] bt

funcrole_t_ROLE_BITTESTANDCOMPLEMENT

< [lock] btc

funcrole_t_ROLE_BITTESTANDRESET

< [lock] btr

funcrole_t_ROLE_BITTESTANDSET

< [lock] bts

funcrole_t_ROLE_BSWAP

< bswap() function (any size)

funcrole_t_ROLE_BUG

< BUG() helper macro: never returns, causes exception

funcrole_t_ROLE_CFSUB3

< carry flag after subtract with carry

funcrole_t_ROLE_CONTAINING_RECORD

< CONTAINING_RECORD() macro

funcrole_t_ROLE_EMPTY

< empty, does not do anything (maybe spoils regs)

funcrole_t_ROLE_FASTFAIL

< __fastfail()

funcrole_t_ROLE_IS_MUL_OK

< is_mul_ok

funcrole_t_ROLE_MEMCPY

< memcpy(void *dst, const void *src, size_t count);

funcrole_t_ROLE_MEMSET

< memset(void *dst, uchar value, size_t count);

funcrole_t_ROLE_MEMSET32

< memset32(void *dst, uint32 value, size_t count);

funcrole_t_ROLE_MEMSET64

< memset64(void *dst, uint64 value, size_t count);

funcrole_t_ROLE_OFSUB3

< overflow flag after subtract with carry

funcrole_t_ROLE_PRESENT

< present() function (used in patterns)

funcrole_t_ROLE_READFLAGS

< __readeflags, __readcallersflags

funcrole_t_ROLE_ROL

< rotate left

funcrole_t_ROLE_ROR

< rotate right

funcrole_t_ROLE_SATURATED_MUL

< saturated_mul

funcrole_t_ROLE_SSE_CMP4

< e.g. _mm_cmpgt_ss

funcrole_t_ROLE_SSE_CMP8

< e.g. _mm_cmpgt_sd

funcrole_t_ROLE_STRCAT

< strcat(char *dst, const char *src);

funcrole_t_ROLE_STRCPY

< strcpy(char *dst, const char *src);

funcrole_t_ROLE_STRLEN

< strlen(const char *src);

funcrole_t_ROLE_TAIL

< char *tail(const char *str);

funcrole_t_ROLE_UNK

< unknown function role

funcrole_t_ROLE_VA_ARG

< va_arg() macro

funcrole_t_ROLE_VA_COPY

< va_copy() function

funcrole_t_ROLE_VA_END

< va_end() function

funcrole_t_ROLE_VA_START

< va_start() function

funcrole_t_ROLE_WCSCAT

< wchar_t *wcscat(wchar_t *dst, const wchar_t *src)

funcrole_t_ROLE_WCSCPY

< wchar_t *wcscpy(wchar_t *dst, const wchar_t *src);

funcrole_t_ROLE_WCSLEN

< size_t wcslen(const wchar_t *s)

funcrole_t_ROLE_WMEMCPY

< wchar_t *wmemcpy(wchar_t *dst, const wchar_t *src, size_t n)

funcrole_t_ROLE_WMEMSET

< wchar_t *wmemset(wchar_t *dst, wchar_t wc, size_t n)

item_preciser_t_ITP_ARG1

< , (64 entries are reserved for 64 call arguments)

item_preciser_t_ITP_ARG64

item_preciser_t_ITP_ASM

< __asm-line

item_preciser_t_ITP_BLOCK1

< opening block comment. this comment is printed before the item < (other comments are indented and printed after the item)

item_preciser_t_ITP_BLOCK2

< closing block comment.

item_preciser_t_ITP_BRACE1

item_preciser_t_ITP_BRACE2

< )

item_preciser_t_ITP_CASE

< bit for switch cases

item_preciser_t_ITP_COLON

< : (label)

item_preciser_t_ITP_CURLY1

< {

item_preciser_t_ITP_CURLY2

< }

item_preciser_t_ITP_DO

< do-line

item_preciser_t_ITP_ELSE

< else-line

item_preciser_t_ITP_EMPTY

< nothing

item_preciser_t_ITP_INNER_LAST

item_preciser_t_ITP_SEMI

< semicolon

item_preciser_t_ITP_SIGN

< if this bit is set too, then we have a negative case value

item_preciser_t_ITP_TRY

< C++ try statement

mba_maturity_t_MMAT_CALLS

< detected call arguments. see also hxe_calls_done

mba_maturity_t_MMAT_GENERATED

< generated microcode

mba_maturity_t_MMAT_GLBOPT1

< performed the first pass of global optimization

mba_maturity_t_MMAT_GLBOPT2

< most global optimization passes are done

mba_maturity_t_MMAT_GLBOPT3

< completed all global optimization. microcode is fixed now.

mba_maturity_t_MMAT_LOCOPT

< local optimization of each basic block is complete. < control flow graph is ready too.

mba_maturity_t_MMAT_LVARS

< allocated local variables

mba_maturity_t_MMAT_PREOPTIMIZED

< preoptimized pass is complete

mba_maturity_t_MMAT_ZERO

< microcode does not exist

mblock_type_t_BLT_0WAY

< does not have successors (tail is a noret function)

mblock_type_t_BLT_1WAY

< passes execution to one block (regular or goto block)

mblock_type_t_BLT_2WAY

< passes execution to two blocks (conditional jump)

mblock_type_t_BLT_NONE

< unknown block type

mblock_type_t_BLT_NWAY

< passes execution to many blocks (switch idiom)

mblock_type_t_BLT_STOP

< stops execution regularly (must be the last block)

mblock_type_t_BLT_XTRN

< external block (out of function address)

mcode_t_m_add

mcode_t_m_and

mcode_t_m_bnot

mcode_t_m_call

mcode_t_m_cfadd

mcode_t_m_cfshl

mcode_t_m_cfshr

mcode_t_m_ext

mcode_t_m_f2f

mcode_t_m_f2i

mcode_t_m_f2u

mcode_t_m_fadd

mcode_t_m_fdiv

mcode_t_m_fmul

mcode_t_m_fneg

mcode_t_m_fsub

mcode_t_m_goto

mcode_t_m_high

mcode_t_m_i2f

mcode_t_m_icall

mcode_t_m_ijmp

mcode_t_m_ja

mcode_t_m_jae

mcode_t_m_jb

mcode_t_m_jbe

mcode_t_m_jcnd

mcode_t_m_jg

mcode_t_m_jge

mcode_t_m_jl

mcode_t_m_jle

mcode_t_m_jnz

mcode_t_m_jtbl

mcode_t_m_jz

mcode_t_m_ldc

mcode_t_m_ldx

mcode_t_m_lnot

mcode_t_m_low

mcode_t_m_mov

mcode_t_m_mul

mcode_t_m_neg

mcode_t_m_nop

mcode_t_m_ofadd

mcode_t_m_or

mcode_t_m_pop

mcode_t_m_push

mcode_t_m_ret

mcode_t_m_sar

mcode_t_m_sdiv

mcode_t_m_seta

mcode_t_m_setae

mcode_t_m_setb

mcode_t_m_setbe

mcode_t_m_setg

mcode_t_m_setge

mcode_t_m_setl

mcode_t_m_setle

mcode_t_m_setnz

mcode_t_m_seto

mcode_t_m_setp

mcode_t_m_sets

mcode_t_m_setz

mcode_t_m_shl

mcode_t_m_shr

mcode_t_m_smod

mcode_t_m_stx

mcode_t_m_sub

mcode_t_m_u2f

mcode_t_m_udiv

mcode_t_m_umod

mcode_t_m_und

mcode_t_m_xds

mcode_t_m_xdu

mcode_t_m_xor

tinfo_t_bitmask_cvt_stance_t_ENUMBM_AUTO

< convert to bitmask if the outcome is nice and useful

tinfo_t_bitmask_cvt_stance_t_ENUMBM_OFF

< convert to ordinal enum

tinfo_t_bitmask_cvt_stance_t_ENUMBM_ON

< convert to bitmask enum

tinfo_t_gta_pdata_t_GTP_BIT_BUCKETS

< get bit buckets

tinfo_t_gta_pdata_t_GTP_COMMENT

< get type comment

tinfo_t_gta_pdata_t_GTP_FINAL_NAME

< get final referenced name

tinfo_t_gta_pdata_t_GTP_NAME

< get referenced name

tinfo_t_gta_pdata_t_GTP_NEXT_NAME

< get immediately next referenced name

tinfo_t_gta_pdata_t_GTP_NICE_NAME

< get the referenced name and apply regular expressions to beautify the name

tinfo_t_gta_pdata_t_GTP_RPTCMT

< get repeatable type comment

tinfo_t_gta_pdata_t_GTP_TIL

< get type library

tinfo_t_gta_pdata_t_GTP_UDT_METHODS

< get udt member functions

tinfo_t_gta_prop_t_GTA_ALIAS

< get type alias

tinfo_t_gta_prop_t_GTA_ARRAY_ELEM

< array: array element

tinfo_t_gta_prop_t_GTA_ARRAY_NELEMS

< array: number of elements

tinfo_t_gta_prop_t_GTA_BITMASK

< enum: is bitmask or regular enum \ref enum_type_data_t::is_bf()

tinfo_t_gta_prop_t_GTA_DECLALIGN

< declared alignment

tinfo_t_gta_prop_t_GTA_EDM

< enum: get enum type member by index

tinfo_t_gta_prop_t_GTA_EDM_BYNAME

< enum: find enum type member by name

tinfo_t_gta_prop_t_GTA_EDM_BYVAL

< enum: find enum type member by value/serial/bmask

tinfo_t_gta_prop_t_GTA_EDM_TID

< enum: get enum member tid

tinfo_t_gta_prop_t_GTA_EDT_NMEMBERS

< enum: get number of enum members

tinfo_t_gta_prop_t_GTA_ENUM_BASE_TYPE

< enum: get enum base type

tinfo_t_gta_prop_t_GTA_ENUM_RADIX

< enum: get enum radix \ref enum_type_data_t::get_enum_radix()

tinfo_t_gta_prop_t_GTA_ENUM_REPR

< enum: get enum value representation

tinfo_t_gta_prop_t_GTA_ENUM_WIDTH

< enum: get enum width \ref enum_type_data_t::calc_nbytes()

tinfo_t_gta_prop_t_GTA_FINAL_ELEM

< if array, skip possible arrays, return a non-array type

tinfo_t_gta_prop_t_GTA_FINAL_ORDINAL

< get final (resolved) type ordinal

tinfo_t_gta_prop_t_GTA_FORWARD_TYPE

< if a forward declaration, return either BTMT_STRUCT/UNION/ENUM

tinfo_t_gta_prop_t_GTA_FRAME_FUNC

< frame: get function address for the frame

tinfo_t_gta_prop_t_GTA_FROM_SUBTIL

< is from a subtil (not from main til)

tinfo_t_gta_prop_t_GTA_FUNC_ARG

< func: get type of function arg

tinfo_t_gta_prop_t_GTA_FUNC_CC

< func: calling convention

tinfo_t_gta_prop_t_GTA_FUNC_NARGS

< func: number of arguments

tinfo_t_gta_prop_t_GTA_FUNC_RET

< func: get function return type

tinfo_t_gta_prop_t_GTA_HAS_UNION

< has members of type “union”?

tinfo_t_gta_prop_t_GTA_HAS_VFTABLE

< has vftable?

tinfo_t_gta_prop_t_GTA_IS_ANON_UDT

< is anonymous struct/union?

tinfo_t_gta_prop_t_GTA_IS_FIXED

< udt: has fixed member offsets?

tinfo_t_gta_prop_t_GTA_IS_FORWARD

< is forward declaration?

tinfo_t_gta_prop_t_GTA_IS_FUNCPTR

< is a pointer to a function?

tinfo_t_gta_prop_t_GTA_IS_HIGH_TYPE

< func: is high type

tinfo_t_gta_prop_t_GTA_IS_SHIFTED_PTR

< is a shifted pointer?

tinfo_t_gta_prop_t_GTA_IS_SMALL_UDT

< udt: is small udt (can be passed in regs)

tinfo_t_gta_prop_t_GTA_IS_SSE_TYPE

< is a SSE vector type?

tinfo_t_gta_prop_t_GTA_IS_TYPEDEF

< is a typedef?

tinfo_t_gta_prop_t_GTA_IS_VARMEMBER

< is a variable member type?

tinfo_t_gta_prop_t_GTA_IS_VARSTRUCT

< is a variable-size structure?

tinfo_t_gta_prop_t_GTA_IS_VFTABLE

< is vftable?

tinfo_t_gta_prop_t_GTA_LAST_FUNC_ARG

tinfo_t_gta_prop_t_GTA_ONEMEM_TYPE

< udt&array: object consisting of one member: type of the member

tinfo_t_gta_prop_t_GTA_ORDINAL

< get initial type ordinal

tinfo_t_gta_prop_t_GTA_PTRARR_SIZE

< ptr&array: get size of subtype

tinfo_t_gta_prop_t_GTA_PTRARR_SUBTIF

< ptr&array: pointed object or array element

tinfo_t_gta_prop_t_GTA_PTR_OBJ

< ptr: pointed type

tinfo_t_gta_prop_t_GTA_PURGED_BYTES

< func: number of purged bytes

tinfo_t_gta_prop_t_GTA_REALTYPE

< real type (do not fully resolve type refs)

tinfo_t_gta_prop_t_GTA_RESOLVE

< real type (fully resolve eventual type references)

tinfo_t_gta_prop_t_GTA_SAFE_PTR_OBJ

< ptr: pointed type or type itself

tinfo_t_gta_prop_t_GTA_TYPE_SIGN

< get type sign

tinfo_t_gta_prop_t_GTA_UDM_IS_BYTIL

< udm: was the member created due to the type system

tinfo_t_gta_prop_t_GTA_UDM_TID

< udt: get member TID

tinfo_t_gta_prop_t_GTA_UDT_NMEMBERS

< udt: get number of udt members

tinfo_t_gta_prop_t_GTA_UNPADDED_SIZE

< udt: sizeof baseclass when embedded into a derived class

tinfo_t_sta_prop_t_STA_ADD_EDM

< enum: add enum member

tinfo_t_sta_prop_t_STA_ADD_FUNCARG

< func: add function argument

tinfo_t_sta_prop_t_STA_ADD_UDM

< udt: add a struct member

tinfo_t_sta_prop_t_STA_ALIAS

< set type alias

tinfo_t_sta_prop_t_STA_ALIGNMENT

< set type alignment

tinfo_t_sta_prop_t_STA_BITMASK

< enum: make/unmake enum a bitmask

tinfo_t_sta_prop_t_STA_CLR_MODIFS

< clear ‘const/volatile’ bits

tinfo_t_sta_prop_t_STA_COMMENT

< set type comment

tinfo_t_sta_prop_t_STA_DECLALIGN

< set declared alignment

tinfo_t_sta_prop_t_STA_DEL_EDMS

< enum: del enum members

tinfo_t_sta_prop_t_STA_DEL_FUNCARGS

< func: del function arguments

tinfo_t_sta_prop_t_STA_DEL_UDMS

< udt: del struct members

tinfo_t_sta_prop_t_STA_EDIT_EDM

< enum: change constant value and/or bitmask

tinfo_t_sta_prop_t_STA_EDM_CMT

< enum: set comment of an enum member

tinfo_t_sta_prop_t_STA_EDM_NAME

< enum: rename enum member

tinfo_t_sta_prop_t_STA_ENUM_RADIX

< enum: set enum radix

tinfo_t_sta_prop_t_STA_ENUM_REPR

< enum: set enum base repr

tinfo_t_sta_prop_t_STA_ENUM_SIGN

< enum: set enum sign

tinfo_t_sta_prop_t_STA_ENUM_WIDTH

< enum: set the width of enum base type

tinfo_t_sta_prop_t_STA_EXPAND_UDT

< udt: expand/shrink struct

tinfo_t_sta_prop_t_STA_FIXED_STRUCT

< struct: use fixed member offsets

tinfo_t_sta_prop_t_STA_FUNCARG_LOC

< func: set argument location

tinfo_t_sta_prop_t_STA_FUNCARG_NAME

< func: rename a function argument

tinfo_t_sta_prop_t_STA_FUNCARG_TYPE

< func: set type of a function argument

tinfo_t_sta_prop_t_STA_FUNC_CC

< func: set calling convention

tinfo_t_sta_prop_t_STA_FUNC_RETLOC

< func: set location of function return value

tinfo_t_sta_prop_t_STA_FUNC_RETTYPE

< func: set function return type

tinfo_t_sta_prop_t_STA_RENAME

< set type name

tinfo_t_sta_prop_t_STA_SET_PACK

< udt: set struct packing

tinfo_t_sta_prop_t_STA_SET_SDA

< udt: set struct alignment

tinfo_t_sta_prop_t_STA_STRUCT_SIZE

< struct: set struct size (only for fixed structs)

tinfo_t_sta_prop_t_STA_TYPE_SIGN

< set type sign

tinfo_t_sta_prop_t_STA_UDM_CMT

< udt: set comment of a struct member

tinfo_t_sta_prop_t_STA_UDM_NAME

< udt: rename a struct member

tinfo_t_sta_prop_t_STA_UDM_REPR

< udt: set repr of a struct member

tinfo_t_sta_prop_t_STA_UDM_SET_BYTIL

< udm: the member is created due to the type system

tinfo_t_sta_prop_t_STA_UDM_TYPE

< udt: set type of a struct member

tinfo_t_sta_prop_t_STA_UDT_ALIGN

< calculate udt field alignments

tinfo_t_sta_prop_t_STA_UDT_METHODS

< set udt member functions

warnid_t_WARN_ADDR_OUTARGS

< 6 cannot handle address arithmetics in outgoing argument area of stack frame – unused

warnid_t_WARN_ARRAY_INARG

< 21 array has been used for an input argument

warnid_t_WARN_BAD_CALL_SP

< 38 bad sp value at call

warnid_t_WARN_BAD_FIELD_TYPE

< 23 incorrect structure member type for %s::%s, ignored

warnid_t_WARN_BAD_INSN

< 49 bad instruction

warnid_t_WARN_BAD_MAPDST

< 48 too short map destination ‘%s’ for variable ‘%s’

warnid_t_WARN_BAD_PURGED

< 12 inconsistent function type and number of purged bytes

warnid_t_WARN_BAD_RETVAR

< 25 wrong return variable

warnid_t_WARN_BAD_SHADOW

< 45 ignored the value written to the shadow area of the succeeding call

warnid_t_WARN_BAD_SP

< 40 positive sp value %a has been found

warnid_t_WARN_BAD_STD_TYPE

< 37 corrupted or unexisting local type ‘%s’

warnid_t_WARN_BAD_STKPNT

< 41 wrong sp change point

warnid_t_WARN_BAD_STROFF

< 33 user specified stroff has not been processed: %s

warnid_t_WARN_BAD_VALRNG

< 44 values range analysis failed

warnid_t_WARN_BAD_VARSIZE

< 34 inconsistent variable size for ‘%s’

warnid_t_WARN_CBUILD_LOOPS

< 13 too many cbuild loops

warnid_t_WARN_CR_BADOFF

< 32 CONTAINING_RECORD: too small offset %d for struct ‘%s’

warnid_t_WARN_CR_NOFIELD

< 31 CONTAINING_RECORD: no field ‘%s’ in struct ‘%s’ at %d

warnid_t_WARN_DEP_UNK_CALLS

< 7 found interdependent unknown calls

warnid_t_WARN_EXP_LINVAR

< 10 failed to expand a linear variable

warnid_t_WARN_FIXED_INSN

< 29 fixed broken insn

warnid_t_WARN_FRAG_LVAR

< 26 fragmented variable at %s may be wrong

warnid_t_WARN_GUESSED_TYPE

< 9 using guessed type %s;

warnid_t_WARN_HUGE_STKOFF

< 27 exceedingly huge offset into the stack frame

warnid_t_WARN_ILL_ELLIPSIS

< 8 erroneously detected ellipsis type has been ignored

warnid_t_WARN_ILL_FPU_STACK

< 18 inconsistent fpu stack

warnid_t_WARN_ILL_FUNCTYPE

< 2 invalid function type ‘%s’ has been ignored

warnid_t_WARN_ILL_PURGED

< 1 odd caller purged bytes %d, correcting

warnid_t_WARN_JUMPOUT

< 43 control flows out of bounds

warnid_t_WARN_MAX

< may be used in notes as a placeholder when the < warning id is not available

warnid_t_WARN_MAX_ARGS

< 22 too many input arguments, some ignored

warnid_t_WARN_MISSED_SWITCH

< 39 wrong markup of switch jump, skipped it

warnid_t_WARN_MUST_RET_FP

< 17 function return type is incorrect (must be floating point)

warnid_t_WARN_NO_SAVE_REST

< 14 could not find valid save-restore pair for %s

warnid_t_WARN_ODD_ABI

< 50 encountered odd instruction for the current ABI

warnid_t_WARN_ODD_ADDR_USE

< 16 odd use of a variable address

warnid_t_WARN_ODD_INPUT_REG

< 15 odd input register %s

warnid_t_WARN_OPT_USELESS_JCND

< 54 simplified comparisons for ‘%s’: %s became %s

warnid_t_WARN_OPT_VALRNG

< 46 conditional instruction was optimized away because %s

warnid_t_WARN_OPT_VALRNG2

< 52 mask 0x%X is shortened because %s <= 0x%X“

warnid_t_WARN_OPT_VALRNG3

< 53 masking with 0X%X was optimized away because %s <= 0x%X

warnid_t_WARN_RET_LOCREF

< 47 returning address of temporary local variable ‘%s’

warnid_t_WARN_SELFREF_PROP

< 19 self-referencing variable has been detected

warnid_t_WARN_SUBFRAME_OVERFLOW

< 55 call arguments overflow the function chunk frame

warnid_t_WARN_UNALIGNED_ARG

< 36 unaligned function argument ‘%s’

warnid_t_WARN_UNBALANCED_STACK

< 51 unbalanced stack, ignored a potential tail call

warnid_t_WARN_UNDEF_LVAR

< 42 variable ‘%s’ is possibly undefined

warnid_t_WARN_UNINITED_REG

< 28 reference to an uninitialized register has been removed: %s

warnid_t_WARN_UNSUPP_REG

< 35 unsupported processor register ‘%s’

warnid_t_WARN_VARARG_MANY

< 5 too many varargs, some ignored

warnid_t_WARN_VARARG_NOSTK

< 4 call vararg without local stack

warnid_t_WARN_VARARG_REGS

< 0 cannot handle register arguments in vararg function, discarded them

warnid_t_WARN_VARARG_TCAL

< 3 cannot handle tail call to vararg

warnid_t_WARN_WIDEN_CHAINS

< 11 failed to widen chains

warnid_t_WARN_WOULD_OVERLAP

< 20 variables would overlap: %s

warnid_t_WARN_WRITE_CONST

< 24 write access to const memory at %a has been detected

warnid_t_WARN_WRONG_VA_OFF

< 30 wrong offset of va_list variable

Statics

ivl_t_allmem⚠

Functions

argloc_t_dstr⚠

bitset_t_extract⚠

bitset_t_print⚠

carglist_t_print⚠

carglist_t_print1⚠

casm_t_genasm⚠

casm_t_print⚠

ccase_t_set_insn⚠

ccases_t_find_value⚠

cexpr_t_cexpr_t⚠

cexpr_t_find_ptr_or_array⚠

Find the pointer operand. This function returns the pointer operand for binary expressions.

cexpr_t_is_aliasable⚠

Check if the expression if aliasable. Simple registers and non-aliasble stack slots return false.

cexpr_t_is_value_used⚠

Does the PARENT need the expression value

cexpr_t_like_boolean⚠

Does the expression look like a boolean expression? In other words, its possible values are only 0 and 1.

cfunc_t_cfunc_t⚠

cfunc_t_cfunc_t1⚠

cgoto_t_print⚠

ctry_t_print⚠

func_t_compare⚠

ivl_t_print⚠

mba_t_get_stkvar⚠

mba_t_mba_t⚠

mba_t_mba_t1⚠

mblock_t_mblock_t⚠

mblock_t_mblock_t_destructor⚠

rangeset_t_check_move_args⚠

Check if the intended move_chunk() arguments are correct.

rangeset_t_count⚠

Size in bytes

rangeset_t_move_chunk⚠

Subtract the address range (from, from+size) and add the range (to, to+size)

rangeset_t_verify⚠

refinfo_t_compare⚠

regarg_t_compare⚠

value_repr_t_compare⚠

vdloc_t_regoff⚠

Type Aliases

_qstring_const_iterator

_qstring_iterator

adiff_t

argloc_t_biggest_t

argloc_type_t

Specifies the location type of a function argument or a debugger variable - see \ref ALOC_

argpartvec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

asize_t

bgcolor_t

bitset_t_const_iterator

boundaries_t

casevec_t

Vector of case values - see calc_switch_cases()

catchexprs_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

ccatchvec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

cinsn_list_t

Linked list Note: linked list is not movable!

cinsnptrvec_t

Vector of pointers to statements.

citem_pointers_t

Vector of parents

cm_t

color_t

@}

const_iterator

ctree_maturity_t

Ctree maturity level. The level will increase as we switch from one phase of ctree generation to the next one

ctype_t

Ctree item code. At the beginning of this list there are expression codes (cot_…), followed by statement codes (cit_…).

ea_t

eamap_t

eavec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

flags64_t

flags_t

funcrole_t

Function roles. They are used to calculate use/def lists and to recognize functions without using string comparisons.

hexwarns_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

int16

int32

int64

intvec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

item_preciser_t

Comment item preciser. Item preciser is used to assign comments to ctree items A ctree item may have several comments attached to it. For example, an if-statement may have the following comments:

if ( … )    // cmt1
{             // cmt2
}             // cmt3
else          // cmt4
{                     – usually the else block has a separate ea
} 
The first 4 comments will have the same ea. In order to denote the exact line for the comment, we store the item_preciser along with ea.

iterator

ivl_t_inherited

ivlset_t_inherited

ivlset_tpl_bag_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

ivlset_tpl_const_iterator

ivlset_tpl_iterator

mba_maturity_t

Microcode maturity levels

mbitmap_t

mblock_type_t

Basic block types

mcallargs_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

mcode_t

mopt_t

Instruction operand types

mopvec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

mreg_t

qlist_const_iterator_value_type

qlist_iterator_value_type

qlist_value_type

qstring

Reimplementation of the string class from STL. Only the most essential functions are implemented. The reason why we have this is because it is not compiler dependent (hopefully) and therefore can be used in IDA API

qvector_const_iterator

qvector_iterator

qvector_value_type

rangeset_t_const_iterator

rangeset_t_iterator

rangevec_base_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

reginfovec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

std_allocator_const_pointer

std_allocator_const_reference

std_allocator_difference_type

std_allocator_is_always_equal

std_allocator_pointer

std_allocator_propagate_on_container_move_assignment

std_allocator_rebind_other

std_allocator_reference

std_allocator_size_type

std_allocator_value_type

std_map__Alloc_traits

std_map__Pair_alloc_type

std_map__Rep_type

std_map_allocator_type

std_map_const_iterator

std_map_const_pointer

std_map_const_reference

std_map_const_reverse_iterator

std_map_difference_type

std_map_insert_return_type

std_map_iterator

std_map_key_compare

std_map_key_type

std_map_mapped_type

std_map_node_type

std_map_pointer

std_map_reference

std_map_reverse_iterator

std_map_size_type

std_map_value_type

std_pair__PCCFP

std_pair__PCCP

std_pair_first_type

std_pair_second_type

strvec_t

A collection of simple lines to populate a custom view. This is an example of what you would pass as the ‘ud’ argument to create_custom_viewer()

sval_t

svalvec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

tinfo_t_bitmask_cvt_stance_t

Set or clear the ‘bitmask’ attribute of an enum. This attribute controls if the enum is considered as a collection of bits or a plain enum. Enums having the ‘bitmask’ attribute can be used to represent bitwise combination of the defined enum members. \param stance \ref bitmask_cvt_stance_t \param etf_flags \ref etf_flag_t

tinfo_t_gta_pdata_t

tinfo_t_gta_prop_t

Various type properties (properties are 64-bit scalar values)

tinfo_t_sta_prop_t

type_attrs_t

this vector must be sorted by keys

type_t

@}

typid_t

@}

uchar

uint8

uint16

uint32

uint64

uint64vec_t

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

user_cmts_t

user_iflags_t

user_labels_t

user_numforms_t

user_unions_t

ushort

uval_ivl_ivlset_t

Set of address intervals. Bit arrays are efficient only for small sets. Potentially huge sets, like memory ranges, require another representation. ivlset_t is used for a list of memory locations in our decompiler.

uval_ivl_t

uval_t

warnid_t

Warning ids

Unions

argloc_t__bindgen_ty_1

cexpr_t__bindgen_ty_1

cexpr_t__bindgen_ty_1__bindgen_ty_1__bindgen_ty_1

cexpr_t__bindgen_ty_1__bindgen_ty_2__bindgen_ty_1

cexpr_t__bindgen_ty_1__bindgen_ty_2__bindgen_ty_2

cinsn_t__bindgen_ty_1

func_t__bindgen_ty_1

mop_t__bindgen_ty_1

The following union holds additional details about the operand. Depending on the operand type different kinds of info are stored. You should access these fields only after verifying the operand type. All pointers are owned by the operand and are freed by its destructor.

value_repr_t__bindgen_ty_1

Additional info