idalib_sys::hexrays::__impl
Struct mba_t

#[repr(C)]
pub struct mba_t {

    flags: c_uint,
    flags2: c_uint,
    pub mbr: mba_ranges_t,
    pub entry_ea: c_ulonglong,
    pub last_prolog_ea: c_ulonglong,
    pub first_epilog_ea: c_ulonglong,
    pub qty: c_int,
    pub npurged: c_int,
    pub cc: c_uchar,
    pub tmpstk_size: c_longlong,
    pub frsize: c_longlong,
    pub frregs: c_longlong,
    pub fpd: c_longlong,
    pub pfn_flags: c_int,
    pub retsize: c_int,
    pub shadow_args: c_int,
    pub fullsize: c_longlong,
    pub stacksize: c_longlong,
    pub inargoff: c_longlong,
    pub minstkref: c_longlong,
    pub minstkref_ea: c_ulonglong,
    pub minargref: c_longlong,
    pub spd_adjust: c_longlong,
    pub gotoff_stkvars: ivlset_t,
    pub restricted_memory: ivlset_t,
    pub aliased_memory: ivlset_t,
    pub nodel_memory: mlist_t,
    pub consumed_argregs: rlist_t,
    pub maturity: c_uint,
    pub reqmat: c_uint,
    pub final_type: bool,
    pub idb_type: tinfo_t,
    pub idb_spoiled: qvector<reg_info_t>,
    pub spoiled_list: mlist_t,
    pub fti_flags: c_int,
    pub label: _qstring<c_char>,
    pub vars: lvars_t,
    pub argidx: qvector<c_int>,
    pub retvaridx: c_int,
    pub error_ea: c_ulonglong,
    pub error_strarg: _qstring<c_char>,
    pub blocks: *mut mblock_t,
    pub natural: *mut *mut mblock_t,
    pub std_ivls: [ivl_with_name_t; 6],
    pub notes: qvector<hexwarn_t>,
    pub occurred_warns: [c_uchar; 32],
    pub reserved: __IncompleteArrayField<c_char>,

}

Array of micro blocks representing microcode for a decompiled function. The first micro block is the entry point, the last one is the exit point. The entry and exit blocks are always empty. The exit block is generated at MMAT_LOCOPT maturity level.

Fields
flags: c_uintflags2: c_uintmbr: mba_ranges_tentry_ea: c_ulonglonglast_prolog_ea: c_ulonglongfirst_epilog_ea: c_ulonglongqty: c_int

< number of basic blocks

npurged: c_int

< -1 - unknown

cc: c_uchar

< calling convention

tmpstk_size: c_longlong

< size of the temporary stack part < (which dynamically changes with push/pops)

frsize: c_longlong

< size of local stkvars range in the stack frame

frregs: c_longlong

< size of saved registers range in the stack frame

fpd: c_longlong

< frame pointer delta

pfn_flags: c_int

< copy of func_t::flags

retsize: c_int

< size of return address in the stack frame

shadow_args: c_int

< size of shadow argument area

fullsize: c_longlong

< Full stack size including incoming args

stacksize: c_longlong

< The maximal size of the function stack including < bytes allocated for outgoing call arguments < (up to retaddr)

inargoff: c_longlong

< offset of the first stack argument; < after fix_scattered_movs() INARGOFF may < be less than STACKSIZE

minstkref: c_longlong

< The lowest stack location whose address was taken

minstkref_ea: c_ulonglong

< address with lowest minstkref (for debugging)

minargref: c_longlong

< The lowest stack argument location whose address was taken < This location and locations above it can be aliased < It controls locations >= inargoff-shadow_args

spd_adjust: c_longlong

< If sp>0, the max positive sp value

gotoff_stkvars: ivlset_t

< stkvars that hold .got offsets. considered to be unaliasable

restricted_memory: ivlset_taliased_memory: ivlset_t

< aliased_memory+restricted_memory=ALLMEM

nodel_memory: mlist_t

< global dead elimination may not delete references to this area

consumed_argregs: rlist_t

< registers converted into stack arguments, should not be used as arguments

maturity: c_uint

< current maturity level

reqmat: c_uint

< required maturity level

final_type: bool

< is the function type final? (specified by the user)

idb_type: tinfo_t

< function type as retrieved from the database

idb_spoiled: qvector<reg_info_t>

< MBA_SPLINFO && final_type: info in ida format

spoiled_list: mlist_t

< MBA_SPLINFO && !final_type: info in vd format

fti_flags: c_int

< FTI_… constants for the current function

label: _qstring<c_char>

< name of the function or pattern (colored)

vars: lvars_t

< local variables

argidx: qvector<c_int>

< input arguments (indexes into ‘vars’)

retvaridx: c_int

< index of variable holding the return value < -1 means none

error_ea: c_ulonglong

< during microcode generation holds ins.ea

error_strarg: _qstring<c_char>blocks: *mut mblock_t

< double linked list of blocks

natural: *mut *mut mblock_t

< natural order of blocks

std_ivls: [ivl_with_name_t; 6]

< we treat memory as consisting of 6 parts < see \ref memreg_index_t

notes: qvector<hexwarn_t>occurred_warns: [c_uchar; 32]reserved: __IncompleteArrayField<c_char>
Implementations

impl mba_t

pub unsafe fn get_stkvar( &self, udm: *mut udm_t, vd_stkoff: c_longlong, p_idaoff: *mut c_ulonglong, p_frame: *mut tinfo_t, ) -> isize

pub unsafe fn new() -> Self