idalib_sys::hexrays::__impl
Type Alias catchexprs_t

pub type catchexprs_t = qvector<catchexpr_t>;

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

Aliased Type

struct catchexprs_t {
    pub _phantom_0: PhantomData<UnsafeCell<catchexpr_t>>,
    array: *mut catchexpr_t,
    n: usize,
    alloc: usize,
}

Fields
_phantom_0: PhantomData<UnsafeCell<catchexpr_t>>array: *mut catchexpr_tn: usizealloc: usize