idalib_sys::hexrays::__impl
Type Alias svalvec_t

pub type svalvec_t = qvector<c_longlong>;

Reimplementation of vector class from STL. Only the most essential functions are implemented. \n The vector container accepts objects agnostic to their positions in the memory because it will move them arbitrarily (realloc and memmove). \n The reason why we have it is because it is not compiler dependent (hopefully) and therefore can be used in IDA API.

Aliased Type

struct svalvec_t {
    pub _phantom_0: PhantomData<UnsafeCell<i64>>,
    array: *mut i64,
    n: usize,
    alloc: usize,
}

Fields
_phantom_0: PhantomData<UnsafeCell<i64>>array: *mut i64n: usizealloc: usize