idalib_sys::hexrays::__impl
Type Alias tinfo_t_bitmask_cvt_stance_t

pub type tinfo_t_bitmask_cvt_stance_t = c_uint;

Set or clear the ‘bitmask’ attribute of an enum. This attribute controls if the enum is considered as a collection of bits or a plain enum. Enums having the ‘bitmask’ attribute can be used to represent bitwise combination of the defined enum members. \param stance \ref bitmask_cvt_stance_t \param etf_flags \ref etf_flag_t

Each group starts with a mask member. Group size is the number of enum constants in it, including group mask. GROUP_SIZES contains the group sizes. Sum of GROUPS_SIZES is equal to number of enum constants. If value is the only value in a group, no need for additional mask value. \note

If mask candidate is equal to the next constant then this is not a bitmask enum

Constant 0 could not be a mask