Struct idalib_sys::hexrays::citem_t

#[repr(C)]
pub struct citem_t {
    pub ea: c_ulonglong,
    pub op: c_uint,
    pub label_num: c_int,
    pub index: c_int,
}

Basic ctree item. This is an abstract class (but we don’t use virtual functions in ctree, so the compiler will not disallow you to create citem_t instances). However, items of pure citem_t type must never be created. Two classes, cexpr_t and cinsn_t are derived from it.

Fields
ea: c_ulonglong

< address that corresponds to the item. may be BADADDR

op: c_uint

< item type

label_num: c_int

< label number. -1 means no label. items of the expression < types (cot_…) should not have labels at the final maturity < level, but at the intermediate levels any ctree item < may have a label. Labels must be unique. Usually < they correspond to the basic block numbers.

index: c_int

< an index in cfunc_t::treeitems. < meaningful only after print_func()