idalib_sys::loader

Function find_plugin

pub unsafe fn find_plugin(
    name: *const c_char,
    load_if_needed: bool,
) -> *mut plugin_t

Find a user-defined plugin and optionally load it. \param name short plugin name without path and extension, or absolute path to the file name \param load_if_needed if the plugin is not present in the memory, try to load it \return pointer to plugin description block