Module idalib_sys::pod

Structs

insn_t

Structure to hold information about an instruction. \ingroup instruction This structure is filled by the analysis step of IDP and used by the emulation and conversion to text steps. The kernel uses this structure too. All structure fields except #cs, #ip, #ea, and op_t::n,op_t::flags of #ops are initialized to zero by the kernel. The rest should be filled by ana().

op_t

Operand of an instruction. \ingroup operands_t This structure is filled by the analyzer. Upon entrance to the analyzer, some fields of this structure are initialized:

op_t__bindgen_ty_2__bindgen_ty_1

This structure is defined for your convenience only

op_t__bindgen_ty_3__bindgen_ty_1

This structure is defined for your convenience only

op_t__bindgen_ty_4__bindgen_ty_1

This structure is defined for your convenience only

Constants

OF_NO_BASE_DISP

OF_NUMBER

OF_OUTER_DISP

OF_SHOW

Type Aliases

asize_t

ea_t

int16

op_dtype_t

optype_t

\defgroup operands Operands Work with instruction operands

uchar

uint8

uint16

uint32

uint64

uval_t

Unions

insn_t__bindgen_ty_1

op_t__bindgen_ty_1

@}

op_t__bindgen_ty_2

op_t__bindgen_ty_3

op_t__bindgen_ty_4