Module idalib_sys::segment

Structs

lock_segment

Helper class to lock a segment pointer so it stays valid

segment_t

Describes a program segment

Constants

SEGPERM_EXEC

SEGPERM_MAXVAL

SEGPERM_READ

SEGPERM_WRITE

SEG_ABSSYM

SEG_BSS

SEG_CODE

SEG_COMM

SEG_DATA

SEG_GRP

SEG_IMEM

SEG_IMP

SEG_MAX_SEGTYPE_CODE

SEG_NORM

SEG_NULL

SEG_UNDF

SEG_XTRN

saAbs

saGroup

saRel4K

saRel32Bytes

saRel64Bytes

saRel128Bytes

saRel512Bytes

saRel1024Bytes

saRel2048Bytes

saRelByte

saRelDble

saRelPage

saRelPara

saRelQword

saRelWord

saRel_MAX_ALIGN_CODE

Functions

get_segm_by_name⚠

Get pointer to segment by its name. If there are several segments with the same name, returns the first of them. \param name segment name. may be nullptr. \return nullptr or pointer to segment structure

get_segm_qty⚠

Get number of segments

getnseg⚠

Get pointer to segment by its number. \warning Obsoleted because it can slow down the debugger (it has to refresh the whole memory segmentation to calculate the correct answer) \param n segment number in the range (0..get_segm_qty()-1) \returns nullptr or pointer to segment structure

getseg⚠

Get pointer to segment by linear address. \param ea linear address belonging to the segment \return nullptr or pointer to segment structure

idalib_segm_align⚠

idalib_segm_bitness⚠

idalib_segm_bytes⚠

idalib_segm_name⚠

idalib_segm_perm⚠

idalib_segm_type⚠