Struct idalib_sys::segment::segment_t

#[repr(C, align(8))]
pub struct segment_t { /* private fields */ }

Describes a program segment

Implementations

impl segment_t

pub unsafe fn compare(self: &segment_t, r: &segment_t) -> c_int

pub unsafe fn new() -> impl New<Output = Self>

Constructor

impl segment_t

pub unsafe fn is_16bit(&self) -> bool

Is a 16-bit segment?

impl segment_t

pub unsafe fn is_32bit(&self) -> bool

Is a 32-bit segment?

impl segment_t

pub unsafe fn is_64bit(&self) -> bool

Is a 64-bit segment?

impl segment_t

pub unsafe fn abits(&self) -> c_int

Get number of address bits

impl segment_t

pub unsafe fn abytes(&self) -> c_int

Get number of address bytes

impl segment_t

pub unsafe fn comorg(&self) -> bool

\name Segment flag: ORG directive See #SFL_COMORG @{

impl segment_t

pub unsafe fn set_comorg(self: Pin<&mut Self>)

impl segment_t

pub unsafe fn clr_comorg(self: Pin<&mut Self>)

impl segment_t

pub unsafe fn ob_ok(&self) -> bool

\name Segment flag: orgbase See #SFL_OBOK @{

impl segment_t

pub unsafe fn set_ob_ok(self: Pin<&mut Self>)

impl segment_t

pub unsafe fn clr_ob_ok(self: Pin<&mut Self>)

impl segment_t

pub unsafe fn is_visible_segm(&self) -> bool

\name Segment flag: hidden See #SFL_HIDDEN @{

impl segment_t

pub unsafe fn set_visible_segm(self: Pin<&mut Self>, visible: bool)

impl segment_t

pub unsafe fn is_debugger_segm(&self) -> bool

\name Segment flag: debugger segment See #SFL_DEBUG @{

impl segment_t

pub unsafe fn set_debugger_segm(self: Pin<&mut Self>, debseg: bool)

impl segment_t

pub unsafe fn is_loader_segm(&self) -> bool

\name Segment flag: loader segment See #SFL_LOADER @{

impl segment_t

pub unsafe fn set_loader_segm(self: Pin<&mut Self>, ldrseg: bool)

impl segment_t

pub unsafe fn is_hidden_segtype(&self) -> bool

\name Segment flag: hide segment type See #SFL_HIDETYPE @{

impl segment_t

pub unsafe fn set_hidden_segtype(self: Pin<&mut Self>, hide: bool)

impl segment_t

pub unsafe fn is_header_segm(&self) -> bool

\name Segment flag: header segment See #SFL_HEADER @{

impl segment_t

pub unsafe fn set_header_segm(self: Pin<&mut Self>, on: bool)

impl segment_t

pub unsafe fn is_ephemeral_segm(&self) -> bool

Ephemeral segments are not analyzed automatically (no flirt, no functions unless required, etc). Most likely these segments will be destroyed at the end of the debugging session unless the user changes their status.

impl segment_t

pub unsafe fn update(self: Pin<&mut Self>) -> bool

Update segment information. You must call this function after modification of segment characteristics. Note that not all fields of segment structure may be modified directly, there are special functions to modify some fields. \return success