idalib_sys

Struct range_t

#[repr(C)]
pub struct range_t {
    pub start_ea: ea_t,
    pub end_ea: ea_t,
}

\file lines.hpp

\brief High level functions that deal with the generation of the disassembled text lines.

This file also contains definitions for the syntax highlighting.

Finally there are functions that deal with anterior/posterior user-defined lines.

Fields
start_ea: ea_t

< start_ea included

end_ea: ea_t

< end_ea excluded

Implementations

impl range_t

pub unsafe fn new(ea1: c_ulonglong, ea2: c_ulonglong) -> impl New<Output = Self>

impl range_t

pub unsafe fn compare(&self, r: &range_t) -> c_int

Compare two range_t instances, based on the start_ea

impl range_t

pub unsafe fn contains(&self, ea: c_ulonglong) -> bool

Is ‘ea’ in the address range?

impl range_t

pub unsafe fn contains1(&self, r: &range_t) -> bool

Is every ea in ‘r’ also in this range_t?

impl range_t

pub unsafe fn overlaps(&self, r: &range_t) -> bool

Is there an ea in ‘r’ that is also in this range_t?

impl range_t

pub unsafe fn clear(self: Pin<&mut Self>)

Set #start_ea, #end_ea to 0

impl range_t

pub unsafe fn empty(&self) -> bool

Is the size of the range_t <= 0?

impl range_t

pub unsafe fn size(&self) -> c_ulonglong

Get #end_ea - #start_ea

impl range_t

pub unsafe fn intersect(self: Pin<&mut Self>, r: &range_t)

Assign the range_t to the intersection between the range_t and ‘r’

impl range_t

pub unsafe fn extend(self: Pin<&mut Self>, ea: c_ulonglong)

Ensure that the range_t includes ‘ea’

impl range_t

pub unsafe fn print(&self, buf: *mut c_char, bufsize: usize) -> usize

Print the range_t. \param buf the output buffer \param bufsize the size of the buffer