idalib_sys

Module func

Modules

cfg_flags

flags

Structs

func_t

A function is a set of continuous ranges of addresses with characteristics

gdl_graph_t

gdl graph interface - includes only functions required to draw it

lock_func

Helper class to lock a function pointer so it stays valid

qbasic_block_t

Information about a basic block of a \ref qflow_chart_t

qflow_chart_t

A flow chart for a function, or a set of address ranges

Enums

fc_block_type_t

Flow chart block types

Functions

calc_thunk_func_target⚠

Calculate target of a thunk function. \param pfn pointer to function (may not be nullptr) \param fptr out: will hold address of a function pointer (if indirect jump) \return the target function or #BADADDR

get_func⚠

Get pointer to function structure by address. \param ea any address in a function \return ptr to a function or nullptr. This function returns a function entry chunk.

get_func_num⚠

Get ordinal number of a function. \param ea any address in the function \return number of function (0..get_func_qty()-1). -1 means ‘no function at the specified address’.

get_func_qty⚠

Get total number of functions in the program

getn_func⚠

Get pointer to function structure by number. \param n number of function, is in range 0..get_func_qty()-1 \return ptr to a function or nullptr. This function returns a function entry chunk.

idalib_func_flags⚠

idalib_func_flow_chart⚠

idalib_func_name⚠

idalib_get_func_cmt⚠

idalib_qbasic_block_preds⚠

idalib_qbasic_block_succs⚠

idalib_qflow_graph_getn_block⚠