idalib_sys

Module loader

Modules

flags

Structs

plugin_t

A plugin is a module in the plugins subdirectory that can perform an action asked by the user. (usually via pressing a hotkey)

Functions

find_plugin⚠

Find a user-defined plugin and optionally load it. \param name short plugin name without path and extension, or absolute path to the file name \param load_if_needed if the plugin is not present in the memory, try to load it \return pointer to plugin description block

idalib_plugin_flags⚠

idalib_plugin_version⚠

run_plugin⚠

Run a loaded plugin with the specified argument. \param ptr pointer to plugin description block \param arg argument to run with