idalib_sys::processor
Struct processor_t

#[repr(C, align(8))]
pub struct processor_t {
    _pinned: PhantomData<PhantomPinned>,
    _non_send_sync: PhantomData<[*const u8; 0]>,
    _data: UnsafeCell<MaybeUninit<[u8; 144]>>,
}

Describes a processor module (IDP). An IDP file may have only one such structure called LPH. The kernel will copy it to dbctx_t::ph.

Fields
_pinned: PhantomData<PhantomPinned>_non_send_sync: PhantomData<[*const u8; 0]>_data: UnsafeCell<MaybeUninit<[u8; 144]>>
Implementations

impl processor_t

fn processor_t_init_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn term_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn newprc_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn newasm_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn asm_installed_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn newfile_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn oldfile_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn newbinary_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn endbinary_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn creating_segm_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn assemble_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn ana_insn_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn emu_insn_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_header_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_footer_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_segstart_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_segend_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_assumes_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_insn_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_out_mnem_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_operand_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_out_data_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_label_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn out_special_item_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn gen_stkvar_def_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn gen_regvar_def_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn gen_src_file_lnnum_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_rename_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn may_show_sreg_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn coagulate_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

pub unsafe fn auto_queue_empty(type_: c_int)

fn func_bounds_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn may_be_func_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_sane_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn cmp_operands_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_jump_func_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_is_basic_block_end_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn getreg_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn undefine_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn moving_segm_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_sp_based_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_far_jump_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_is_call_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_is_ret_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_align_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn can_have_type_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_stkvar_scale_factor_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn demangle_name_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn create_flat_group_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_alloca_probe_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_reg_name_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn gen_asm_or_lst_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn gen_map_file_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_autocmt_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn loader_elf_machine(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: Problem handling function argument p_procname: Pointer pointed to another pointer, which is not yet supported

fn is_indirect_jump_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn verify_noreturn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn verify_sp_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn create_func_frame_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_frame_retsize_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn analyze_prolog_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_spdelta_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calcrel_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_reg_accesses_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_control_flow_guard_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn find_reg_value_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn find_op_value_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn treat_hindering_item_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn extract_address_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn str2reg_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_switch_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn create_switch_xrefs_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_switch_cases_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_bg_color_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn validate_flirt_func_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_operand_string_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_add_cref_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn processor_t_add_dref_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn del_cref_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn del_dref_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn coagulate_dref_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

pub unsafe fn set_idp_options( keyword: *const c_char, vtype: c_int, value: *const c_void, idb_loaded: bool, ) -> *const c_char

fn set_proc_options_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn adjust_libfunc_ea_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

pub unsafe fn realcvt( m: *mut c_void, e: *mut fpvalue_t, swt: c_ushort, ) -> fpvalue_error_t

fn adjust_refinfo_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn is_cond_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn set_code16_mode_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

pub unsafe fn get_code16_mode(ea: c_ulonglong) -> bool

fn next_exec_insn_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: processor_t_regval_getter_t

fn calc_step_over_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_macro_insn_head_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_dbr_opnum_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn insn_reads_tbit_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: processor_t_regval_getter_t

fn get_idd_opinfo_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: processor_t_regval_getter_t

fn calc_next_eas_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn clean_tbit_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: processor_t_regval_getter_t

pub unsafe fn get_reg_info( regname: *const c_char, bitrange: *mut bitrange_t, ) -> *const c_char

Get register information - useful for registers like al, ah, dil, etc. Example: this function for “al” returns “eax” in 32bit mode \return main register name (nullptr no such register)

fn update_call_stack_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item depends on some other type(s) which autocxx could not generate, some of them are: processor_t_regval_getter_t

fn setup_til_autocxx_wrapper_0xdde2bac051107010(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn max_ptr_size_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_cdecl_purged_bytes_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn equal_reglocs_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn decorate_name_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_retloc_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_varglocs_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_arglocs_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn use_stkarg_type_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn use_regarg_type_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn use_arg_types_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn calc_purged_bytes_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_stkarg_area_info_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_cc_regs_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_simd_types_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn arg_addrs_ready_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn adjust_argloc_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn lower_func_type_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn get_abi_info_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn arch_changed_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn create_merge_handlers_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn privrange_changed_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn cvt64_supval_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

fn cvt64_hashval_autocxx_wrapper_0xdde2bac051107010( _uhoh: BindingGenerationFailure, )

autocxx bindings couldn’t be generated: This item relies on a type not known to autocxx (isize)

pub unsafe fn get_regfinder() -> *mut reg_finder_t

pub unsafe fn is_canon_insn(self: &processor_t, itype: c_ushort) -> bool

Does the given value specify a valid instruction for this instruction set?. See #instruc_start and #instruc_end

pub unsafe fn get_canon_mnem( self: &processor_t, itype: c_ushort, ) -> *const c_char

Get the instruction name

pub unsafe fn get_canon_feature(self: &processor_t, itype: c_ushort) -> c_uint

Get the instruction features (combination of \ref CF_)

fn ensure_processor(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This function was marked =delete

fn is_funcarg_off(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This function was marked =delete

fn lvar_off(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This function was marked =delete

fn is_lumina_usable(_uhoh: BindingGenerationFailure)

autocxx bindings couldn’t be generated: This function was marked =delete

impl processor_t

pub unsafe fn has_idp_opts(&self) -> bool

impl processor_t

pub unsafe fn has_segregs(&self) -> bool

impl processor_t

pub unsafe fn use32(&self) -> bool

impl processor_t

pub unsafe fn use64(&self) -> bool

impl processor_t

pub unsafe fn ti(&self) -> bool

impl processor_t

pub unsafe fn stkup(&self) -> bool

impl processor_t

pub unsafe fn use_tbyte(&self) -> bool

impl processor_t

pub unsafe fn use_mappings(&self) -> bool

impl processor_t

pub unsafe fn has_code16_bit(&self) -> bool

impl processor_t

pub unsafe fn supports_macros(&self) -> bool

impl processor_t

pub unsafe fn supports_calcrel(&self) -> bool

impl processor_t

pub unsafe fn calcrel_in_bits(&self) -> bool

impl processor_t

pub unsafe fn get_default_segm_bitness(&self, is_64bit_app: bool) -> c_int

Get default segment bitness \retval 2 #PR_DEFSEG64 \retval 1 #PR_DEFSEG32 \retval 0 none specified

impl processor_t

pub unsafe fn cbsize(self: Pin<&mut Self>) -> c_int

\name Byte size Number of 8bit bytes required to hold one byte of the target processor. @{

impl processor_t

pub unsafe fn dbsize(self: Pin<&mut Self>) -> c_int

impl processor_t

pub unsafe fn get_proc_index(self: Pin<&mut Self>) -> c_int

@}

impl processor_t

pub unsafe fn delay_slot_insn( self: Pin<&mut Self>, ea: *mut c_ulonglong, bexec: *mut bool, fexec: *mut bool, ) -> bool

Helper function to get the delay slot instruction

impl processor_t

pub unsafe fn get_stkvar_scale(self: Pin<&mut Self>) -> c_int

Get the stack variable scaling factor. Useful for processors who refer to the stack with implicit scaling factor. TMS320C55 for example: SP(#1) really refers to (SP+2)

impl processor_t

pub unsafe fn sizeof_ldbl(&self) -> usize

Get size of long double