idalib_sys

Module util

Functions

is_align_insn⚠

If the instruction at ‘ea’ looks like an alignment instruction, return its length in bytes. Otherwise return 0.

is_basic_block_end⚠

Is the instruction the end of a basic block?

is_call_insn⚠

Is the instruction a “call”?

is_indirect_jump_insn⚠

Is the instruction an indirect jump?

is_ret_insn⚠

@}

next_head⚠

Get start of next defined item. \param ea begin search at this address \param maxea not included in the search range \return #BADADDR if none exists.

prev_head⚠

Get start of previous defined item. \param ea begin search at this address \param minea included in the search range \return #BADADDR if none exists.

str2reg⚠